refactor: 재사용 가능한 뷰 분리
